multi now includes decimal

+++ src/main/java/com/insurgencedev/deluxesellwandsaddon/listeners/DeluxeSellwandsEventListener.java
@@ -30,6 +30,6 @@ public final class DeluxeSellwandsEventListener implements Listener {
     }
 
     private double calculateAmount(double amount, double multi) {
-        return (long) (amount * (multi < 1 ? 1 + multi : multi));
+        return amount * (multi < 1 ? 1 + multi : multi);
     }
 }
